IA-Service/requirements.txt
pydantic
aio_pika
python-dotenv
groq